Moved comps to common folder
