+++ package.json
@@ -5,7 +5,8 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "start": "next start"
+    "start": "next start",
+    "lint": "npx prettier components/**/*.js --write"
   },
   "dependencies": {
     "next": "9.4.1",
